fix(schedule): wire schedule runs into edge routing and CLI contract boundary
- fleet-edge-runtime routes the schedule-runs read action alongside list/show
- fleet-command-route compares host-guard paths via canonicalRoot
- cli declares cron-parser runtime dep so packaged smoke resolves it
- allow schedule-runs-contract in the thin CLI static import boundary
- refresh command-parity and gui bridge fixtures for the new read
