+++ packages/cli/package.json
@@ -40,6 +40,7 @@
     "@effect/platform": "0.96.2",
     "@effect/sql": "0.51.1",
     "@effect/sql-sqlite-node": "0.52.0",
+    "cron-parser": "5.10.0",
     "effect": "3.21.4",
     "node-pty": "1.1.0"
   }
