+++ packages/daemon/test/command-admission.test.ts
@@ -10,8 +10,8 @@ const localSource = "local" as const;
 const assignmentSource = { kind: "assignment", nodeId: "edge-one", assignmentId: "assignment-one" } as const;
 const admissionRoutes = new Set(["direct", "via-assignment", "via-center-forward", "rejected"]);
 
-test("all 137 daemon commands close every repo-mode admission cell", () => {
-  assert.equal(daemonProtocolCommands.length, 137);
+test("all 138 daemon commands close every repo-mode admission cell", () => {
+  assert.equal(daemonProtocolCommands.length, 138);
   let cells = 0;
   for (const command of daemonProtocolCommands) {
     assert.deepEqual(Object.keys(command.admission).sort(), [...daemonRepoModeWords].sort(), command.id);
@@ -35,7 +35,7 @@ test("all 137 daemon commands close every repo-mode admission cell", () => {
       }
     }
   }
-  assert.equal(cells, 137 * 3);
+  assert.equal(cells, 138 * 3);
 });
 
 test("observe.tail declares direct admission and named source residency for every tail kind", () => {
@@ -63,7 +63,7 @@ test("Schedule descriptors derive all three mode routes without a CLI mode branc
       "remote-center": "via-assignment",
       "remote-edge": "via-center-forward",
     });
-  for (const id of ["schedule-list", "schedule-show"])
+  for (const id of ["schedule-list", "schedule-runs", "schedule-show"])
     assert.deepEqual(byId.get(id)?.admission, {
       local: "direct",
       "remote-center": "direct",
