+++ packages/daemon/test/json-rpc-protocol.test.ts
@@ -279,8 +279,8 @@ test("GUI action facets are exact, typed, and exclude the generic runner", () =>
     ["repo.agentRuntime.spawn", { runtimeInstanceId: "instance-codex", cwd: { scope: "repo-root" }, prompt: "Inspect", taskId: null, idempotencyKey: "runtime-once" }],
     ["repo.agent.entity.write", { declaration: { schema: "agent-declaration/v1", id: "gui-created-agent", name: "GUI Created Agent", instructions: "Keep the roster intact.\nSecond line.", runtime_type: "any", role: "worker", model: "gpt-5.6-terra", skills: [{ id: "review", path: "skills/review" }], prompts: ["prompt://gui"], preset: "standard-task" } }],
     ["repo.squad.entity.write", { declaration: { schema: "squad-declaration/v1", id: "gui-created-squad", name: "GUI Created Squad", leader: "gui-created-agent", workers: ["gui-created-agent"], leaderTurnBudget: 8, roster: "## GUI Squad\n\n  GUI Created Agent\n\n" } }],
-    ["repo.schedule.create", { scheduleId: "schedule-a", name: "Schedule A", everyMs: 300000, agentId: "agent-a", runtimeInstanceId: "instance-a", mission: "Run A.", idempotencyKey: "schedule-create-once" }],
-    ["repo.schedule.update", { scheduleId: "schedule-a", name: "Updated A", everyMs: 600000, agentId: "agent-a", runtimeInstanceId: "instance-a", mission: "Run updated A.", model: null, reasoningEffort: null, cwd: null, idempotencyKey: "schedule-update-once" }],
+    ["repo.schedule.create", { scheduleId: "schedule-a", name: "Schedule A", mode: "detect", everyMs: 300000, agentId: "agent-a", runtimeInstanceId: "instance-a", mission: "Run A.", idempotencyKey: "schedule-create-once" }],
+    ["repo.schedule.update", { scheduleId: "schedule-a", name: "Updated A", mode: "remediate", everyMs: 600000, agentId: "agent-a", runtimeInstanceId: "instance-a", mission: "Run updated A.", model: null, reasoningEffort: null, cwd: null, idempotencyKey: "schedule-update-once" }],
     ["repo.schedule.delete", { scheduleId: "schedule-a", reason: "retired", idempotencyKey: "schedule-delete-once" }],
     ["repo.schedule.enable", { scheduleId: "schedule-a", idempotencyKey: "schedule-enable-once" }],
     ["repo.schedule.disable", { scheduleId: "schedule-a", idempotencyKey: "schedule-disable-once" }],
